[Decomp][Unit] Cohesion split bundle: extract faction predicates to UnitHostility.cpp + visibility/detection to UnitVisibility.cpp
