+++ src/game/Object/Unit.cpp
@@ -6448,358 +6448,6 @@ FactionTemplateEntry const* Unit::getFactionTemplateEntry() const
 }
 
 /**
- * @brief Checks whether this unit is hostile to another unit.
- *
- * @param unit The other unit.
- * @return True if the units are hostile; otherwise, false.
- */
-bool Unit::IsHostileTo(Unit const* unit) const
-{
-    // always non-hostile to self
-    if (unit == this)
-    {
-        return false;
-    }
-
-    // always non-hostile to GM in GM mode
-    if (unit->GetTypeId() == TYPEID_PLAYER && ((Player const*)unit)->isGameMaster())
-    {
-        return false;
-    }
-
-    // always hostile to enemy
-    if (getVictim() == unit || unit->getVictim() == this)
-    {
-        return true;
-    }
-
-    // test pet/charm masters instead pers/charmeds
-    Unit const* testerOwner = GetCharmerOrOwner();
-    Unit const* targetOwner = unit->GetCharmerOrOwner();
-
-    // always hostile to owner's enemy
-    if (testerOwner && (testerOwner->getVictim() == unit || unit->getVictim() == testerOwner))
-    {
-        return true;
-    }
-
-    // always hostile to enemy owner
-    if (targetOwner && (getVictim() == targetOwner || targetOwner->getVictim() == this))
-    {
-        return true;
-    }
-
-    // always hostile to owner of owner's enemy
-    if (testerOwner && targetOwner && (testerOwner->getVictim() == targetOwner || targetOwner->getVictim() == testerOwner))
-    {
-        return true;
-    }
-
-    Unit const* tester = testerOwner ? testerOwner : this;
-    Unit const* target = targetOwner ? targetOwner : unit;
-
-    // always non-hostile to target with common owner, or to owner/pet
-    if (tester == target)
-    {
-        return false;
-    }
-
-    // special cases (Duel, etc)
-    if (tester->GetTypeId() == TYPEID_PLAYER && target->GetTypeId() == TYPEID_PLAYER)
-    {
-        Player const* pTester = (Player const*)tester;
-        Player const* pTarget = (Player const*)target;
-
-        // Duel
-        if (pTester->IsInDuelWith(pTarget))
-        {
-            return true;
-        }
-
-        // Group
-        if (pTester->GetGroup() && pTester->GetGroup() == pTarget->GetGroup())
-        {
-            return false;
-        }
-
-        // Sanctuary
-        if (pTarget->HasFlag(PLAYER_FLAGS, PLAYER_FLAGS_SANCTUARY) && pTester->HasFlag(PLAYER_FLAGS, PLAYER_FLAGS_SANCTUARY))
-        {
-            return false;
-        }
-
-        // PvP FFA state
-        if (pTester->IsFFAPvP() && pTarget->IsFFAPvP())
-        {
-            return true;
-        }
-
-        //= PvP states
-        // Green/Blue (can't attack)
-        if (pTester->GetTeam() == pTarget->GetTeam())
-        {
-            return false;
-        }
-
-        // Red (can attack) if true, Blue/Yellow (can't attack) in another case
-        return pTester->IsPvP() && pTarget->IsPvP();
-    }
-
-    // faction base cases
-    FactionTemplateEntry const* tester_faction = tester->getFactionTemplateEntry();
-    FactionTemplateEntry const* target_faction = target->getFactionTemplateEntry();
-    if (!tester_faction || !target_faction)
-    {
-        return false;
-    }
-
-    if (target->isAttackingPlayer() && tester->IsContestedGuard())
-    {
-        return true;
-    }
-
-    // PvC forced reaction and reputation case
-    if (tester->GetTypeId() == TYPEID_PLAYER)
-    {
-        if (target_faction->faction)
-        {
-            // forced reaction
-            if (ReputationRank const* force = ((Player*)tester)->GetReputationMgr().GetForcedRankIfAny(target_faction))
-            {
-                return *force <= REP_HOSTILE;
-            }
-
-            // if faction have reputation then hostile state for tester at 100% dependent from at_war state
-            if (FactionEntry const* raw_target_faction = sFactionStore.LookupEntry(target_faction->faction))
-                if (FactionState const* factionState = ((Player*)tester)->GetReputationMgr().GetState(raw_target_faction))
-                {
-                    return (factionState->Flags & FACTION_FLAG_AT_WAR);
-                }
-        }
-    }
-    // CvP forced reaction and reputation case
-    else if (target->GetTypeId() == TYPEID_PLAYER)
-    {
-        if (tester_faction->faction)
-        {
-            // forced reaction
-            if (ReputationRank const* force = ((Player*)target)->GetReputationMgr().GetForcedRankIfAny(tester_faction))
-            {
-                return *force <= REP_HOSTILE;
-            }
-
-            // apply reputation state
-            FactionEntry const* raw_tester_faction = sFactionStore.LookupEntry(tester_faction->faction);
-            if (raw_tester_faction && raw_tester_faction->reputationListID >= 0)
-            {
-                return ((Player const*)target)->GetReputationMgr().GetRank(raw_tester_faction) <= REP_HOSTILE;
-            }
-        }
-    }
-
-    // common faction based case (CvC,PvC,CvP)
-    return tester_faction->IsHostileTo(*target_faction);
-}
-
-/**
- * @brief Checks whether this unit is friendly to another unit.
- *
- * @param unit The other unit.
- * @return True if the units are friendly; otherwise, false.
- */
-bool Unit::IsFriendlyTo(Unit const* unit) const
-{
-    // always friendly to self
-    if (unit == this)
-    {
-        return true;
-    }
-
-    // always friendly to GM in GM mode
-    if (unit->GetTypeId() == TYPEID_PLAYER && ((Player const*)unit)->isGameMaster())
-    {
-        return true;
-    }
-
-    // always non-friendly to enemy
-    if (getVictim() == unit || unit->getVictim() == this)
-    {
-        return false;
-    }
-
-    // test pet/charm masters instead pers/charmeds
-    Unit const* testerOwner = GetCharmerOrOwner();
-    Unit const* targetOwner = unit->GetCharmerOrOwner();
-
-    // always non-friendly to owner's enemy
-    if (testerOwner && (testerOwner->getVictim() == unit || unit->getVictim() == testerOwner))
-    {
-        return false;
-    }
-
-    // always non-friendly to enemy owner
-    if (targetOwner && (getVictim() == targetOwner || targetOwner->getVictim() == this))
-    {
-        return false;
-    }
-
-    // always non-friendly to owner of owner's enemy
-    if (testerOwner && targetOwner && (testerOwner->getVictim() == targetOwner || targetOwner->getVictim() == testerOwner))
-    {
-        return false;
-    }
-
-    Unit const* tester = testerOwner ? testerOwner : this;
-    Unit const* target = targetOwner ? targetOwner : unit;
-
-    // always friendly to target with common owner, or to owner/pet
-    if (tester == target)
-    {
-        return true;
-    }
-
-    // special cases (Duel)
-    if (tester->GetTypeId() == TYPEID_PLAYER && target->GetTypeId() == TYPEID_PLAYER)
-    {
-        Player const* pTester = (Player const*)tester;
-        Player const* pTarget = (Player const*)target;
-
-        // Duel
-        if (pTester->IsInDuelWith(pTarget))
-        {
-            return false;
-        }
-
-        // Group
-        if (pTester->GetGroup() && pTester->GetGroup() == pTarget->GetGroup())
-        {
-            return true;
-        }
-
-        // Sanctuary
-        if (pTarget->HasFlag(PLAYER_FLAGS, PLAYER_FLAGS_SANCTUARY) && pTester->HasFlag(PLAYER_FLAGS, PLAYER_FLAGS_SANCTUARY))
-        {
-            return true;
-        }
-
-        // PvP FFA state
-        if (pTester->IsFFAPvP() && pTarget->IsFFAPvP())
-        {
-            return false;
-        }
-
-        //= PvP states
-        // Green/Blue (non-attackable)
-        if (pTester->GetTeam() == pTarget->GetTeam())
-        {
-            return true;
-        }
-
-        // Blue (friendly/non-attackable) if not PVP, or Yellow/Red in another case (attackable)
-        return !pTarget->IsPvP();
-    }
-
-    // faction base cases
-    FactionTemplateEntry const* tester_faction = tester->getFactionTemplateEntry();
-    FactionTemplateEntry const* target_faction = target->getFactionTemplateEntry();
-    if (!tester_faction || !target_faction)
-    {
-        return false;
-    }
-
-    if (target->isAttackingPlayer() && tester->IsContestedGuard())
-    {
-        return false;
-    }
-
-    // PvC forced reaction and reputation case
-    if (tester->GetTypeId() == TYPEID_PLAYER)
-    {
-        if (target_faction->faction)
-        {
-            // forced reaction
-            if (ReputationRank const* force = ((Player*)tester)->GetReputationMgr().GetForcedRankIfAny(target_faction))
-            {
-                return *force >= REP_FRIENDLY;
-            }
-
-            // if faction have reputation then friendly state for tester at 100% dependent from at_war state
-            if (FactionEntry const* raw_target_faction = sFactionStore.LookupEntry(target_faction->faction))
-                if (FactionState const* factionState = ((Player*)tester)->GetReputationMgr().GetState(raw_target_faction))
-                {
-                    return !(factionState->Flags & FACTION_FLAG_AT_WAR);
-                }
-        }
-    }
-    // CvP forced reaction and reputation case
-    else if (target->GetTypeId() == TYPEID_PLAYER)
-    {
-        if (tester_faction->faction)
-        {
-            // forced reaction
-            if (ReputationRank const* force = ((Player*)target)->GetReputationMgr().GetForcedRankIfAny(tester_faction))
-            {
-                return *force >= REP_FRIENDLY;
-            }
-
-            // apply reputation state
-            if (FactionEntry const* raw_tester_faction = sFactionStore.LookupEntry(tester_faction->faction))
-                if (raw_tester_faction->reputationListID >= 0)
-                {
-                    return ((Player const*)target)->GetReputationMgr().GetRank(raw_tester_faction) >= REP_FRIENDLY;
-                }
-        }
-    }
-
-    // common faction based case (CvC,PvC,CvP)
-    return tester_faction->IsFriendlyTo(*target_faction);
-}
-
-/**
- * @brief Checks whether this unit is intrinsically hostile to players.
- *
- * @return True if the faction template is hostile to players; otherwise, false.
- */
-bool Unit::IsHostileToPlayers() const
-{
-    FactionTemplateEntry const* my_faction = getFactionTemplateEntry();
-    if (!my_faction || !my_faction->faction)
-    {
-        return false;
-    }
-
-    FactionEntry const* raw_faction = sFactionStore.LookupEntry(my_faction->faction);
-    if (raw_faction && raw_faction->reputationListID >= 0)
-    {
-        return false;
-    }
-
-    return my_faction->IsHostileToPlayers();
-}
-
-/**
- * @brief Checks whether this unit is neutral to all factions.
- *
- * @return True if the faction template is neutral to all; otherwise, false.
- */
-bool Unit::IsNeutralToAll() const
-{
-    FactionTemplateEntry const* my_faction = getFactionTemplateEntry();
-    if (!my_faction || !my_faction->faction)
-    {
-        return true;
-    }
-
-    FactionEntry const* raw_faction = sFactionStore.LookupEntry(my_faction->faction);
-    if (raw_faction && raw_faction->reputationListID >= 0)
-    {
-        return false;
-    }
-
-    return my_faction->IsNeutralToAll();
-}
-
-/**
  * @brief Starts attacking a victim.
  *
  * @param victim The victim to attack.
@@ -8346,382 +7994,6 @@ int32 Unit::ModifyPower(Powers power, int32 dVal)
     }
 
     return gain;
-}
-
-/**
- * @brief Checks whether the unit is visible to or detectable by another unit.
- *
- * @param u The observing unit.
- * @param viewPoint The viewpoint used for checks.
- * @param detect True to include stealth detection logic.
- * @param inVisibleList True when evaluating retention in an existing visible list.
- * @param is3dDistance True to use 3D distance checks.
- * @return True if the unit should be seen or detected; otherwise, false.
- */
-bool Unit::IsVisibleForOrDetect(Unit const* u, WorldObject const* viewPoint, bool detect, bool inVisibleList, bool is3dDistance) const
-{
-    if (!u || !IsInMap(u))
-    {
-        return false;
-    }
-
-    // Always can see self
-    if (u == this)
-    {
-        return true;
-    }
-
-    // player visible for other player if not logout and at same transport
-    // including case when player is out of world
-    bool at_same_transport =
-        GetTypeId() == TYPEID_PLAYER &&  u->GetTypeId() == TYPEID_PLAYER &&
-        !((Player*)this)->GetSession()->PlayerLogout() && !((Player*)u)->GetSession()->PlayerLogout() &&
-        !((Player*)this)->GetSession()->PlayerLoading() && !((Player*)u)->GetSession()->PlayerLoading() &&
-        ((Player*)this)->GetTransport() && ((Player*)this)->GetTransport() == ((Player*)u)->GetTransport();
-
-    // not in world
-    if (!at_same_transport && (!IsInWorld() || !u->IsInWorld()))
-    {
-        return false;
-    }
-
-    // forbidden to seen (while Removing corpse)
-    if (m_Visibility == VISIBILITY_REMOVE_CORPSE)
-    {
-        return false;
-    }
-
-    Map& _map = *u->GetMap();
-    // Grid dead/alive checks
-    if (u->GetTypeId() == TYPEID_PLAYER)
-    {
-        // non visible at grid for any stealth state
-        if (!IsVisibleInGridForPlayer((Player*)u))
-        {
-            return false;
-        }
-
-        // if player is dead then he can't detect anyone in any cases
-        if (!u->IsAlive())
-        {
-            detect = false;
-        }
-    }
-    else
-    {
-        // all dead creatures/players not visible for any creatures
-        if (!u->IsAlive() || !IsAlive())
-        {
-            return false;
-        }
-    }
-
-    // different visible distance checks
-    if (u->IsTaxiFlying())                                  // what see player in flight
-    {
-        // use object grey distance for all (only see objects any way)
-        if (!IsWithinDistInMap(viewPoint, World::GetMaxVisibleDistanceInFlight() + (inVisibleList ? World::GetVisibleObjectGreyDistance() : 0.0f), is3dDistance))
-        {
-            return false;
-        }
-    }
-    else if (!at_same_transport)                            // distance for show player/pet/creature (no transport case)
-    {
-        // Any units far than max visible distance for viewer or not in our map are not visible too
-        if (!IsWithinDistInMap(viewPoint, _map.GetVisibilityDistance() + (inVisibleList ? World::GetVisibleUnitGreyDistance() : 0.0f), is3dDistance))
-        {
-            return false;
-        }
-    }
-
-    // always seen by owner
-    if (GetCharmerOrOwnerGuid() == u->GetObjectGuid())
-    {
-        return true;
-    }
-
-    // isInvisibleForAlive() those units can only be seen by dead or if other
-    // unit is also invisible for alive.. if an isinvisibleforalive unit dies we
-    // should be able to see it too
-    if (u->IsAlive() && IsAlive() && IsInvisibleForAlive() != u->IsInvisibleForAlive())
-    {
-        if (u->GetTypeId() != TYPEID_PLAYER || !((Player*)u)->isGameMaster())
-        {
-            return false;
-        }
-    }
-
-    // Visible units, always are visible for all units, except for units under invisibility and phases
-    if (m_Visibility == VISIBILITY_ON && u->m_invisibilityMask == 0)
-    {
-        return true;
-    }
-
-    // GMs see any players, not higher GMs and all units in any phase
-    if (u->GetTypeId() == TYPEID_PLAYER && ((Player*)u)->isGameMaster())
-    {
-        if (GetTypeId() == TYPEID_PLAYER)
-        {
-            return ((Player*)this)->GetSession()->GetSecurity() <= ((Player*)u)->GetSession()->GetSecurity();
-        }
-        else
-        {
-            return true;
-        }
-    }
-
-    // non faction visibility non-breakable for non-GMs
-    if (m_Visibility == VISIBILITY_OFF)
-    {
-        return false;
-    }
-
-    // grouped players should always see stealthed party members
-    if (GetTypeId() == TYPEID_PLAYER && u->GetTypeId() == TYPEID_PLAYER)
-    {
-        if (((Player*)this)->IsGroupVisibleFor(((Player*)u)) && u->IsFriendlyTo(this))
-        {
-            return true;
-        }
-    }
-
-    // raw invisibility
-    bool invisible = (m_invisibilityMask != 0 || u->m_invisibilityMask != 0);
-    if (u->GetTypeId() == TYPEID_PLAYER) // if object is player with mover, use its visibility masks, so that an invisible player MCing a creature can see stuff
-    {
-        if (Player* player = (Player*)u)
-        {
-            if (Unit* mover=player->GetMover())
-            {
-                invisible= (m_invisibilityMask != 0 || mover->m_invisibilityMask != 0);
-            }
-        }
-    }
-
-    // detectable invisibility case
-    if (invisible && (
-                // Invisible units, always are visible for units under same invisibility type
-                (m_invisibilityMask & u->m_invisibilityMask) != 0 ||
-                // Invisible units, always are visible for unit that can detect this invisibility (have appropriate level for detect)
-                u->canDetectInvisibilityOf(this) ||
-                // Units that can detect invisibility always are visible for units that can be detected
-                canDetectInvisibilityOf(u)))
-    {
-        invisible = false;
-    }
-
-    // special cases for always overwrite invisibility/stealth
-    if (invisible || m_Visibility == VISIBILITY_GROUP_STEALTH)
-    {
-        if (u->IsHostileTo(this))
-        {
-            // Hunter mark functionality
-            AuraList const& auras = GetAurasByType(SPELL_AURA_MOD_STALKED);
-            for (AuraList::const_iterator iter = auras.begin(); iter != auras.end(); ++iter)
-                if ((*iter)->GetCasterGuid() == u->GetObjectGuid())
-                {
-                    return true;
-                }
-        }
-
-        // none other cases for detect invisibility, so invisible
-        if (invisible)
-        {
-            return false;
-        }
-    }
-
-    // unit got in stealth in this moment and must ignore old detected state
-    if (m_Visibility == VISIBILITY_GROUP_NO_DETECT)
-    {
-        return false;
-    }
-
-    // GM invisibility checks early, invisibility if any detectable, so if not stealth then visible
-    if (m_Visibility != VISIBILITY_GROUP_STEALTH)
-    {
-        return true;
-    }
-
-    // NOW ONLY STEALTH CASE
-
-    // if in non-detect mode then invisible for unit
-    // mobs always detect players (detect == true)... return 'false' for those mobs which have (detect == false)
-    // players detect players only in Player::HandleStealthedUnitsDetection()
-    if (!detect)
-    {
-        return (u->GetTypeId() == TYPEID_PLAYER) ? ((Player*)u)->HaveAtClient(this) : false;
-    }
-
-    // Special cases
-
-    // If is attacked then stealth is lost, some creature can use stealth too
-    if (!getAttackers().empty())
-    {
-        return true;
-    }
-
-    // If there is collision rogue is seen regardless of level difference
-    if (IsWithinDist(u, 0.24f))
-    {
-        return true;
-    }
-
-    // If a mob or player is stunned he will not be able to detect stealth
-    if (u->hasUnitState(UNIT_STAT_STUNNED) && (u != this))
-    {
-        return false;
-    }
-
-    // set max ditance
-    float visibleDistance = (u->GetTypeId() == TYPEID_PLAYER) ? MAX_PLAYER_STEALTH_DETECT_RANGE : ((Creature const*)u)->GetAttackDistance(this);
-
-    // Always invisible from back (when stealth detection is on), also filter max distance cases
-    bool isInFront = viewPoint->IsInFrontInMap(this, visibleDistance);
-    if (!isInFront)
-    {
-        return false;
-    }
-
-    // if doesn't have stealth detection (Shadow Sight), then check how stealthy the unit is, otherwise just check los
-    if (!u->HasAuraType(SPELL_AURA_DETECT_STEALTH))
-    {
-        // Calculation if target is in front
-
-        // Visible distance based on stealth value (stealth rank 4 300MOD, 10.5 - 3 = 7.5)
-        visibleDistance = 10.5f - (GetTotalAuraModifier(SPELL_AURA_MOD_STEALTH) / 100.0f);
-
-        // Visible distance is modified by
-        //-Level Diff (every level diff = 1.0f in visible distance)
-        visibleDistance += int32(u->GetLevelForTarget(this)) - int32(GetLevelForTarget(u));
-
-        // This allows to check talent tree and will add addition stealth dependent on used points)
-        int32 stealthMod = GetTotalAuraModifier(SPELL_AURA_MOD_STEALTH_LEVEL);
-        if (stealthMod < 0)
-        {
-            stealthMod = 0;
-        }
-
-        //-Stealth Mod(positive like Master of Deception) and Stealth Detection(negative like paranoia)
-        // based on wowwiki every 5 mod we have 1 more level diff in calculation
-        visibleDistance += (int32(u->GetTotalAuraModifier(SPELL_AURA_MOD_STEALTH_DETECT)) - stealthMod) / 5.0f;
-        visibleDistance = visibleDistance > MAX_PLAYER_STEALTH_DETECT_RANGE ? MAX_PLAYER_STEALTH_DETECT_RANGE : visibleDistance;
-
-        // recheck new distance
-        if (visibleDistance <= 0 || !IsWithinDist(viewPoint, visibleDistance))
-        {
-            return false;
-        }
-    }
-
-    // Now check is target visible with LoS
-    float ox, oy, oz;
-    viewPoint->GetPosition(ox, oy, oz);
-    return IsWithinLOS(ox, oy, oz);
-}
-
-/**
- * @brief Refreshes visibility, viewpoint auras, and AI visibility notifications.
- */
-void Unit::UpdateVisibilityAndView()
-{
-
-    static const AuraType auratypes[] = {SPELL_AURA_BIND_SIGHT, SPELL_AURA_FAR_SIGHT, SPELL_AURA_NONE};
-    for (AuraType const* type = &auratypes[0]; *type != SPELL_AURA_NONE; ++type)
-    {
-        AuraList& alist = m_modAuras[*type];
-        if (alist.empty())
-        {
-            continue;
-        }
-
-        for (AuraList::iterator it = alist.begin(); it != alist.end();)
-        {
-            Aura* aura = (*it);
-            Unit* owner = aura->GetCaster();
-
-            if (!owner || !IsVisibleForOrDetect(owner, this, false))
-            {
-                alist.erase(it);
-                RemoveAura(aura);
-                it = alist.begin();
-            }
-            else
-            {
-                ++it;
-            }
-        }
-    }
-
-    GetViewPoint().Call_UpdateVisibilityForOwner();
-    UpdateObjectVisibility();
-    ScheduleAINotify(0);
-    GetViewPoint().Event_ViewPointVisibilityChanged();
-}
-
-/**
- * @brief Sets the unit visibility mode and refreshes visibility if needed.
- *
- * @param x The new visibility mode.
- */
-void Unit::SetVisibility(UnitVisibility x)
-{
-    m_Visibility = x;
-
-    if (IsInWorld())
-    {
-        UpdateVisibilityAndView();
-    }
-}
-
-/**
- * @brief Checks whether this unit can detect another unit's invisibility.
- *
- * @param u The potentially invisible unit.
- * @return True if invisibility can be detected; otherwise, false.
- */
-bool Unit::canDetectInvisibilityOf(Unit const* u) const
-{
-    if (uint32 mask = (m_detectInvisibilityMask & u->m_invisibilityMask))
-    {
-        for (int32 i = 0; i < 32; ++i)
-        {
-            if (((1 << i) & mask) == 0)
-            {
-                continue;
-            }
-
-            // find invisibility level
-            int32 invLevel = 0;
-            Unit::AuraList const& iAuras = u->GetAurasByType(SPELL_AURA_MOD_INVISIBILITY);
-            for (Unit::AuraList::const_iterator itr = iAuras.begin(); itr != iAuras.end(); ++itr)
-                if ((*itr)->GetModifier()->m_miscvalue == i && invLevel < (*itr)->GetModifier()->m_amount)
-                {
-                    invLevel = (*itr)->GetModifier()->m_amount;
-                }
-
-            // find invisibility detect level
-            int32 detectLevel = 0;
-            Unit::AuraList const& dAuras = GetAurasByType(SPELL_AURA_MOD_INVISIBILITY_DETECTION);
-            for (Unit::AuraList::const_iterator itr = dAuras.begin(); itr != dAuras.end(); ++itr)
-                if ((*itr)->GetModifier()->m_miscvalue == i && detectLevel < (*itr)->GetModifier()->m_amount)
-                {
-                    detectLevel = (*itr)->GetModifier()->m_amount;
-                }
-
-            if (i == 6 && GetTypeId() == TYPEID_PLAYER)     // special drunk detection case
-            {
-                detectLevel = ((Player*)this)->GetDrunkValue();
-            }
-
-            if (invLevel <= detectLevel)
-            {
-                return true;
-            }
-        }
-    }
-
-    return false;
 }
 
 /**
